my_tattoo_artist/backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.1
django-cors-headers==4.0.0
django-resized==1.0.2
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.5.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.2
